Use selected text in searchToolBar
Summary:
With this patch, the currently selected text in a webpage is automatically inserted into the
searchToolBar when it is opened.
This matches the behaviour of the search tool bar in Kate.

Test Plan:
Open a webpage. Select a single line of text. Press Ctrl+F. The selected text is inserted into the
searchToolBar.

Reviewers: #falkon, drosca

Reviewed By: #falkon, drosca

Subscribers: drosca

Differential Revision: https://phabricator.kde.org/D11210