FancyTabWidget: Don't draw tab title
This hides tab text in BrowsingLibrary as it would often get elided
due to insufficient space in tab bar. It's not possible to resize
the tab bar to fit the tab text as the text can be very long depending
on translation, so the only way to solve it is to just hide the text.