KWalletPasswords, GnomeKeyringPasswords: Implement migration from QupZilla
Implement migration of passwords stored in KWallet or GNOME Keyring from
QupZilla to Falkon.

The logic is the same as for the profiles: If no "Falkon" folder exists,
but a "QupZilla" folder exists, we read all the data from the "QupZilla"
folder and write it to the "Falkon" folder. (In GNOME Keyring, those are
not actually folders, but application attributes on the entries, but the
logic is the same.)

In GnomeKeyringPasswordBackend, a storeEntry helper method is introduced
to avoid duplicating code.

Reviewed By: drosca
Differential Revision: https://phabricator.kde.org/D12846