Fix crash during download when the main window have been closed
Summary:
Falkon has a feature to automatically close the tab that was specifically
opened for the file to download, because this kind of tab has no useful
purpose to the user.

In the case where the main window would have been closed, Falkon
would crash de-referencing the now NULL pointer to the main window.

This patch adds a simple check to do nothing in this case, as there is no
tab to close anyway.

Reviewers: #falkon, drosca

Reviewed By: #falkon, drosca

Subscribers: drosca, falkon, #falkon

Tags: #falkon

Differential Revision: https://phabricator.kde.org/D14711