Create new bookmark items near the current one (over it)
Summary:
Currently in the bookmarks manager the user can create new items via the contextual menu, so a position is implicitly chosen, but the new items are always inserted at the first position in the list.

This change aims to add new items just before the current (selected) one. If the selected item is a folder the behavior is unchanged (the new item will be inserted at the first position in that folder).

Test Plan:
- Select a separator/url item
- Add a new item (separator/url/folder)
- The new item is inserted just before the previously selected one

Reviewers: #falkon, drosca

Reviewed By: #falkon, drosca

Subscribers: cullmann, falkon

Tags: #falkon

Differential Revision: https://phabricator.kde.org/D14930