VerticalTabs: Corectly recognize tree collapse / expand button
Summary:
In treeview when the tab has no children the collapse / expand button
is not visible but the control logic still thought that is is.
By clicking on place where this button normaly is nothing happens
and user has to click again somewhere else to change the selected tab.

Test Plan:
In treeview mode check the function of collapse / expand button when visible and invisible.

  - Visible: Expand / collapse the tree
  - Not visible: Select tab

Reviewers: #falkon, drosca

Reviewed By: #falkon, drosca

Subscribers: falkon

Tags: #falkon

Differential Revision: https://phabricator.kde.org/D21304