Correct the sorting for cookies white- and black- lists
The previous solution has no effect on entries which already exist in the lists.
As a result the lists are not sorted. It is only functional when the lists were
initially empty, which is a sub-optimal case.

Test plan: open Cookies Manager, switch to Cookie Filtering tab, see the items being alphabetically sorted