Stop installing app icon in pixmaps location
The /usr/share/pixmaps location is considered a legacy location for
application icons; since the application icons are already installed in
the global XDG hicolor theme, then simply stop installing the 128px one
in the legacy pixmaps location.

Copy the 128px icon from the hicolor directory instead, as it is the
same.