Use the correct url when processing the request
The URL can change due to redirect during the same loading request,
which will make the lastLoadingRequest kind of invalid, at least the URL
portion of it.

I still want to keep the old request around but force the URL to be
always updated, maybe I should copy the type and make my own class for
it to handle the changing URL bit more gracefully.

This all could be avoided, if there was an easier way to react on the
network response received signal. Well, maybe in later Qt version or it
is already there an I am ignorant.

Signed-off-by: Juraj Oravec <jurajoravec@mailo.com>